requirements.txt
Flask
google-generativeai
Werkzeug
gunicorn